pyproject.toml
[project]
name = "my-first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.98.0",
    "requests>=2.32.5",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.0",
]
